12.1. Abstract Classes(Section 3: Modern PHP 7 Introduction (OPTIONAL)
In the last lecture we talk about the class inheritance and now it's time to introduce the concept of

abstract class.

Currently none of our classes is abstract and we create free classes.

Jason similar YAML and damp its contents and also use their day or convert Medhat to convert the data

to a specific format.