12.2. Abstract Classes(Section 3: Modern PHP 7 Introduction (OPTIONAL)
But we can actually also create a new instance of the base format class.

First we will have to either use statement for it because otherwise B B won't know where to find this

class.

So either you statement create a new instance of the base format class and remember to pass the data

to its constructor.

Now we can also dump the base class to see the data that it has.

And finally let's call that base format class instance convert Medhat after refreshing the page.

You can see that we have the data inside the base format class.

It's the same as we pass the same data to every one of them and the convert returns.

I am not converting anything.