12.3. Abstract Classes(Section 3: Modern PHP 7 Introduction (OPTIONAL)
Now let's make the base for math class up struct by prepending the class with that abstract keyword

this will make the class base format an abstract class and you cannot instantiate that base class.

The abstract classes directly as you can see the error message you cannot instantiate an abstract class

so abstract classes are created only to be extended by a concrete class that extends this base class.

So let's comment out the class creation and convert method.

And you can see that the upstroke class can have an implementation of some of the methods but it can

also have an abstract method abstract methods can only be added to an abstract class and they can't

have the body.

That means that the abstract method only contains a signature of them method but it cannot actually

have the method body.

The actual body of the meds have to be implemented in the class that extends the base class and it's

also important to note that any class that has at least one abstract method should also be defined as

abstract.

So to recap an abstract class cannot be instantly directly on the abstract class can contain abstract

methods and every class that contains abstract methods has to beat itself abstract and abstract methods

have to be implemented in a class that extends this abstract class.