13.1. Interfaces(Section 3: Modern PHP 7 Introduction (OPTIONAL)
In the last lecture we covered the concept of an abstract class which helped us achieve two things.

First thing was to move some common logic that our Jason SML and yum classes had into a one place that

has had the same code that we would have to otherwise put into every individual class.

But we also made sure using an abstract method that all of these classes actually implement a method

called convert.

Now let's try to add another abstract method into a base format and call it convert from string to achieve

opposite result to convert string to an array.

As you can see we can't instantiate classes that do not have this method.