13.2. Interfaces(Section 3: Modern PHP 7 Introduction (OPTIONAL)
Implement that and there is another way to force classes to contain specific logic or basically to contain

specific methods.

And this is called an interface let's create a new interface using the keyboard interface instead of

the class and call it from string interface interfaces never contain any implementations of the methods.

They only contain their signatures.

So lets the public function convert from stream to from string interface

and.

Now let's try to call method convert from string.

On the Jasen object type some example.

Jason and let's see what will happen.

You might remember that we haven't added convert from string method to any of the classes so we should

expect what should happen.

It should throw an error.

And as expected it says we are calling an undefined method.

Jason convert from string

now we can actually implement that interface inside Jason class using the word implements this speech.

Be that the Class Jason Schrute have the method that is defined inside the from string interface.

So any method signature that you will put inside the from string interface when a certain class implements

that interface it has to implement all the methods that are defined inside this interface.

So now you can see that calling this method works.

It returns no.

But at least it does not throw an error if you will modify the signature of the method inside the interface.

You can see up that declaration inside the adjacent class is different from the interface.

And this is another thing about interfaces.

It has to be exactly the same.

So interfaces are a better method to force some class to contain specific functionality that an abstract

class because sometimes you don't need or do not want every functionality that comes from their abstract

class.

Instead you just want to force the class to contain specific functionality like for example one method

convert from string.

Now you can see that other classes do not implement that interface.

They only have the common convert matter that comes from the base form of abstract class.