13.3. Interfaces(Section 3: Modern PHP 7 Introduction (OPTIONAL)
So now let's see a common example.

When this is very useful to have the classes that you are certain that contain the same functionality

and this would be if you would have array of objects currently you are aware that they are.

Jason ex-MIL Yamhill but consider you might have like a thousand objects are created at random and let's

try to iterate over all these elements by using for each and for each of those formats called that convert

method.

Currently this will work because they are all extending an abstract base for math class that does have

the convert method.

So you can see that there are that there is some Onyx M-L anyway and we call that convert from string

specifically outside the loop for that Jason.

So it works.

Now let's try to make a call to the convert from string.

For every possible format that is out there inside the array and see what happens.

Obviously we have an error because the example class does not implement the convert from string method.

It's not an abstract method in base form up nor the Exxon-Mobile or Jamo class implemented.

They're from string interface.

There is a way we can check if the class is of a certain type or if it implements a certain interface

you can use the keyword instance off if you will surround this call to convert from string with the

if statement.

As you see on the screen it will only called metadata for the class that actually implements that interface.

But if we will dump the result of format instance of from string interface you can see it is false.

And this is because we haven't imported the interface after you use statement for the from string interface.

You can see that the method is called once.

It is called forward Jason.

Class that actually implements that interface and also the format instance of the.

When we dump it it returns true.

You can also use the get class function to get the actual class name of the specific object that is

inside the variable.

So you can see that for each of those class instances that we have inside the array it gives us the

class name of that specific class instance.

Now Peter B does not support multiple inheritance.