13.4. Interfaces(Section 3: Modern PHP 7 Introduction (OPTIONAL)
So you cannot extend more than one class from one particular class and interfaces are the way to force

a certain class to contain some functionality or just make sure that they will contain some specific

functionality that you expect.

And let's create a new class or a new interface call it named form of interface that will have one method

signature which is our Get main

Now add the use statement for this interface inside the index BHB and

you can implement more than one interface.

Impeach B so you can only extend one class but you can implement many interfaces.

Let's also implement the named form format interface in all of the remaining classes.

After adding the implements Q.

Let's see what happens.

The error is that the Jason class contains one abstract method which means it implements interface but

that does not contain the medaled implementation.

So we need to added other good name public met up inside the adjacent class and mind that every signature

every method signature from interface is public.

It cannot be private.

It cannot be protected.

It is always public.

So let's not get name inside XML and YAML and let's copy that statement and check each format from the

array.

If it is of type named format interface and this will actually mean in case of interfaces whether this

class implements that interface.

So if the class if implements named format interface will echoed the get name.

As you can see since all of the classes implement that interface we have Dasent SML and yum output on

the screen.