13.5. Interfaces(Section 3: Modern PHP 7 Introduction (OPTIONAL)
Finally let's try to remove the implements part from the middle class and also remove the get name matter.

It could stator because if it does not implement the interface it's nothing wrong to keep it there but

now you can see how can we check if specific classes that we can have on the array contain specific

functionality so we can safely call it we check if each of the objects contain the get name medaled

by checking if they implement their named format interface so it only calls the get name on objects

that actually implement that particular interface.