14.1. Typed Arguments &amp; Function Return Types(Section 3: Modern PHP 7 Introduction (OPTIONAL)
In this lecture we will talk about the function and class method argument types and also the functions

and methods return types.

So let's you know function call it convert data and convert the function will accept one parameter called

format but this parameter will be typed to base format.

This means that this function expects the format parameter to be of type base format.

And that means that the this should be any of the classes that actually extend that base for my abstract

class.

And that's because their base format is an abstract class.

You cannot instantiate it.

So in this case this would mean that it has to be a class that extends the base form up.

So let's try to damp the result of calling convert data on their Dasent class.

You see that it works both with Jason and with Exham L..