14.2. Typed Arguments &amp; Function Return Types(Section 3: Modern PHP 7 Introduction (OPTIONAL)
But if you will try to instantiate SDD class and pass it.

You see that it has to be an instance of the base form up seems as the class is just a standard empty

class in each B.

And it does not extend the base format.

We have an error.