14.3. Typed Arguments &amp; Function Return Types(Section 3: Modern PHP 7 Introduction (OPTIONAL)
Now we can also force.

We can also force that parameter to a method B of type array.

And since we expected the data field inside base format to be an array adding the array to the method

signature will cause an error if you pass our string.

For example

Now let's revert back to argument.

US data rate.

So it works again.

And now let's try to call the said data method of the Jason class.

We have a string or actually we have to first create that object because it's not created.

I forgot to do that so let's add a new Jason's statement and pass the data as a parameter.

And now let's pass the Jason class instance to convert data.

So this works up to the line 18 where you can see that we made it accept only parameters that are on

array and we have passed a string.