14.4. Typed Arguments &amp; Function Return Types(Section 3: Modern PHP 7 Introduction (OPTIONAL)
We can also set default values for the parameters.

So if you will not pass on parameter to the constructor it would by default use an empty array.

So first we do not have a default parameter to the constructor.

So you see we have an error that at least one argument was expected but let's add the default value.

And after refreshing you can see that we still have an error because we are passing string to the method

that expects an array.

But now it works.

The constructor expects an array.

If you don't provide one.

It uses a default one.