14.5. Typed Arguments &amp; Function Return Types(Section 3: Modern PHP 7 Introduction (OPTIONAL)
Now let's see an example.

When we define a return type of function let's create a function get form out name it will accept an

argument of the type named format interface which means this arguments should implement that interface

and will return the value of the get named function.

So let's try to damp the result of this function called on the adjacent object.

And this is Jason because JSM class implements this interface and returns the string.

Jason.

Now let's try to modify it to number for example number one.

And let's see what would happen.

After refreshing you can see that it returned number one but it doesn't make sense because we expect

a name.

So if you specify a return type as a string it converted the number to a string 1.

But if you declare Street types equals one.

Now it has strict type checks and if the return value was expected to be of type string and you return

a number array or any other type it would show an error.