14.6. Typed Arguments &amp; Function Return Types(Section 3: Modern PHP 7 Introduction (OPTIONAL)
We can also at the same return type to that get name and now it only works if you will return a string.

So that's a very good way to be sure what type you can expect and also make sure that you are returning

the type that you actually expected from that method.

Now it's important to know that you will have to add this Declare statement in every file to make it

work.

You cannot just add it once you have to put it into every single file.

Now let's see what happens if we change the return type of their interface and implement and method

would not have a return type.

It causes an error because their signature often metalled from the interface and the method implemented

in class has to be exactly the same.

And this also applies to the return type declaration.