14.7. Typed Arguments &amp; Function Return Types(Section 3: Modern PHP 7 Introduction (OPTIONAL)
Now let's the new function and call it good form by name and this function will accept our name of the

format that you expect.

So we will be trying to find a class instance that has the get name method returning the name that we

expect and we will specify the return type or actually shouldn't be named for my interface because we

are looking for our class that can convert something.

So let's use base for map and this is how you can make written type option.

So this will be valid for anything that we return a class that extends based format.

But you can also return no

so by using the question mark before the actual type.

You can make it nullable nullable.

It means that it is valid to return only the variable or data of the type specified.

But also you can return.

Now let's create format's array w contain the instances of our free classes the JSM SML and yum.

Let's also the parameter to get the former by name which would be of type array.

And inside this function we will iterate over every element of that array.

We will check if those classes implement the named form format interface and if they do we will also

check if their get named function returns what we expect inside the name parameter.

Then we will return the format object to the one that we were looking for.

But if nothing would be found we will return.

No by default.

So the way it works is that it will iterate over all of them.

And don't forget to re-enabled the get name function inside the inside the Exxon-Mobile class and also

to update the method signatures so it matches the interface signature.

And now let's try to call the get form a by name by looking for

some format.

So first we have to pass the array of formats and you can see that it was found because Jayde some implements

name format interface.

And the get name starts returned Jason.

Now we could not find example and this is because we removed the named format interface from the list

of implement that interfaces by that class.

So it does not contain the get name class so it wasn't even called because we made sure if we first

made sure that the class contains it.