14.8. Typed Arguments &amp; Function Return Types(Section 3: Modern PHP 7 Introduction (OPTIONAL)
Now if we will try to find example you can see that there is a fatal error.

If we removed the question mark from the return type because now it is not optional and it has defined

class of base form.

So you can use it if you are pretty certain that certain classes should be found or when the parameter

could be either something or no.