14.10. Typed Arguments &amp; Function Return Types(Section 3: Modern PHP 7 Introduction (OPTIONAL)
Now let's try to call this method using first element of formats which would be JSM and it works fine

because it just calls the convert method.

But now if you try to return the result of convert.

You can see that there is an error or because of void function must not return a value.