15.1. Anonymous Functions (Closures)(Section 3: Modern PHP 7 Introduction (OPTIONAL)
In this lecture we will cover the anonymous function's anonymous function is serving the purpose which

the callbacks served back in the days.

But the callbacks have to be a function that you define by name.

So you have to write a function provide a name for the function arguments and like Typically when you

define a function and with an anonymous function you can just write a function that will only exist

for example as an argument that you will direct the boss to some other function arguments.

For example the array filter which filters an array to find elements that you expect expect a second

argument to be COBOL or collabo can be a function that you defined traditionally.

Or we can do it as you see on the screen by just inline creating a function that just doesn't have a

name does not exist anywhere outside this context where it is used.

And for example the Arrive filter on the format's array when the anonymous function just checks if that

format or every element of the array has a name.