15.2. Anonymous Functions (Closures)(Section 3: Modern PHP 7 Introduction (OPTIONAL)
For example in this case example IT JUS.

It is the anonymous functions are just very useful for cases like that.

So previously you always had to define function by name.

It had to live somewhere.

And now whenever you need to do something in specific context you can just use anonymous functions.

So let's modified that find by name function

and this function will try to find format that from the list of formats which is an array.

It will try to find a class that would contain the name that you pass to the parameters.

So if the write filter will return nothing we will just return no.

But if there is at least one element because Count will return the number of elements if it would return

0 then it is equal to false return.

No by default.

So let's call the find by name using SML and our formats array.

And let's see what happens.