15.3. Anonymous Functions (Closures)(Section 3: Modern PHP 7 Introduction (OPTIONAL)
And you can see that there was undefined variable name because to use an variable from outer scope from

like outside the anonymous function you have to add another statement after the function signature which

is the use statement.

And here you have to define the variable name from the outside scope that you would like to use.

So you can see that we have found that ESMO.

It also works for Jason because there is a class that returns Jason.

But if you will type something else like existing It will return.

No.

And this is also thanks to our return type bank level.

So to recap previously we will just add on your function with have name for example find by name and

pass it to the array filter.

And in this example we were able to define a function in line that only lives in this particular chunk

of code and cannot be reused everywhere else.

So for cases like that anonymous functions are perfect.