16. Reflection API (Reverse-Engineer Classes, Methods, Functions)(Section 3: Modern PHP 7 Introduction (OPTIONAL)
In this lecture we will talk about cold reflections.

It's not exactly a very modern feature.

It is currently very extensively used in all of the BHP frameworks.

So let's just start straight away and create new reflection class instance and pass the Jason class

name to it.

So it is special.

Can't stand that every class has and it does it will just return the fully qualified class name so you

can see that reflection class itself does not tell us a lot about the Jason class.

It only returns the name which we already knew because it's what is contained inside the class Constand

of Jason class.

But it also has some merit.

It's like class get constructor and it returns an object refl. method which contains the name of the

constructor and the class where it is actually defined.

So in case of the Jason class scenes it extends Beys format and does not have an owner constructor that

constructor leaves inside the base format.

Now the mettled variable is of type reflection method which itself has its own methods.

So you can for example call the parameters.

Well we can just dump the parameters and you can see that it is on the array of the reflection parameter

objects and it only holds the name of the parameter which you can see was data.

Now to find find out more than just the name of the parameters let's iterate over all of the parameters

of that specific method.

In our case it's only one parameter but if we will have more we can iterate over them.

And first let's try to just dump the parameter get type which should return the type of parameter you

can see that it's an object thus does not contain any fields.

So let's return that into a new type variable and we will try to dump it and typecasts it to stream.

You might remember that we did this already and it will just convert the contents of an object to stream.

Because it implemented to stream method.

But there is an interesting method is built in.

So you can see that the type is array and is built in means if it comes from the PH P-word or whether

this is a class or type that was defined by you.

So obviously the array is a part of PKP.

Now let's see if it allows.

No that's another method of that parameter.

It does not allow.

No because it is not an nullable parameter.

We can also check

if it contains a default value.

So you can see that it actually does.

And it is an empty array.

So it just returns any default value.

So if you will modify it to string type and set a default value as a streaming you will see that type

in streaming that it is also a built in type that it does not allow.

No and that the default value is hello.

Now let's also make this parameter Lebel and refresh this.

And now you see after refreshing it also allows move value because we made it Newble and reflections

are very useful to figure out a lot about the code and it really helps them modern P2P frameworks work

great.

But you will see how that works in a couple of the next lectures when we will implement a pretty complicated

thing called servies container and outtalk wiring.

It's obviously a feature of their Symphonie framework but we will implement that.

So you will better understand how that actually works.