17. Dependency Injection (In Practice)(Section 3: Modern PHP 7 Introduction (OPTIONAL))
In this lecture we will cover the dependency injection first let's remove the constructor from the base

for math class because we want to be using it anymore.

That data will be passed using the said data method.

Now we also need to create an interface.

Let's call it our format interface and this interface will be implemented by free of our converting

classes which is the J-Zone example and the Jamo.

Now this interface should have two methods.

One is to set data.

And second would be to convert mettled we create this interface because later we will create Serializer

class which only purpose is to convert that data in the array form to a string.

But we wouldn't really care what would be the format.

We will inject the dependency into this class constructor.

So dependency could be adjacent ecce homo or Yamma class but we wouldn't care which of these classes

that was so Serializer class could do its job.

Which whatever would be passed to it because they all will implement the format interface.

Now since all these classes implement the form of interface we need to add the proper implements statement

and also we have to make sure that all of those methods there said data and convert have exactly the

same signatures as they are in the interface because otherwise we will get an error or

so let's remove the data variable being passed to constructor's of all of these classes.

And let's finally create Serializer class the serializer class will do one thing it will handle a process

that is called serialization and that means that the data that is represented in one way would be converted

to a different representation.

And in our specific case that means that the data that we have as a peach b array would be converted

to either Jason Exxon-Mobile or Jamo.

But the fingers that we would like to make this class generally.

And what I mean by this class being generally is that since it accepts an interface as a dependency

it will work the same for all of the free classes that we have.

Jason X-amount and Yum.

But imagine that in the future you will create some new format whatever that might be.

And then you wouldn't have to change a feeling inside the serializer class itself.

So this class can handle any case because it is just accepting an interface as dependency.

So if you'd like to convert to different formats and you wouldn't use dependency injection but for example

you will pass another parameter or let's say format to the serialize method.

You will probably have to add an if statement.

Double check what format was expected.

And then create an appropriate class by using the dependency injection.

We can just eliminate that process so the code stays the same all the unit tests will work the same.

You will just have to create another class implementing their form format interface and pass it as a

dependency.

So what dependency injection helps us achieve is we eliminate the coupling between the certain parts

of our application.

So whenever you add a new feeling or you change the way things work you wouldn't have to modify a lot

of different classes that rely explicitly on other code like Serializer class does not really know about

the Jason class it just cares that it gets something that can convert data to a different format.

What.

Any class that implements format interface can do.

And in this example we seem to have an error.

And that's because we imported the form and interface from the rank namespace so let's fix that.

And now you see that it is converted to Jason.

If you pass another dependency you can see that it also works for example.

So that's a great benefit of dependency injection that the serializer class in this example but of course

when your apps get bigger you will have a lot of cases like that.

The Serializer class does not really care about the actual class that can do a thing like converting

the arrays to any other representation.

It just needs an interface and it is being passed as a dependency using their constructor.