18.1 Simple Service Container (Hands-On Coding! )(Section 3: Modern PHP 7 Introduction (OPTIONAL))
In the last lecture we talked about dependency injection and how does help us create classes that do

one thing.

And bouse Bustos classes two different classes that can do another thing in the specific example that

we had.

We created Serializer class that can get and other classes which implement the format the interface

to complete the process of serialization.

It's important for you to understand that what we do in this lecture and this whole section tries to

mimic what Symphonie framework does itself.

And basically all the other ph B frameworks out there and even some Java frameworks.

But we do it a much simpler way to help you understand what happens in this modern development world.

So let's create a class and the cold index controller and we will talk later about what controllers

are.

But for now let's just say that controller is like an entry point for a certain address.

Well that just does one thing and returns our response back to the browser user or any other client.

So our index controller will have a constructor that will accept a one dependency which is Serializer

and it will have a one method called index.

The logic behind controllers is very simple you have a cause which is a user enter your browser and

the response is an action.

So the index method here is such an action.

So since we passed the serializer to the index controller we can now use its features.

So you can for example return a response that will be simple array and say that the action is index

and also it will return current time.

So now let's imagine we are on the index page.

So we are creating a new instance of the index controller class which accepts one depends on one dependency

which is a serializer.

So we are passing it to it.

Note that the serializer itself has its own dependencies.

Now what's left is we have to use statement for our index controller class and after refreshing you

can see that we have the data.

The action index and the current time which is being outputted as an example if we change that dependency

of Serializer you can see that now it is Dasen.