18.2 Simple Service Container (Hands-On Coding! )(Section 3: Modern PHP 7 Introduction (OPTIONAL))
Now let's create a new class container.

And this would be our service container.

So services are objects that do some things and the service container as its name suggests is a place

where all of these services are collected so you can fetch them from this container.

And it also organizes the way how all of those services are being created and what dependencies are

being passed to it.

So it is a very convenient way when you will have much more classes that we have in this example.

So this would happen when we will start using the symphonic framework and API platform that we will

need a way to organize all of those services.

Just to be clear because I dont want to scare you service container is already a part of the symphonic

framework.

So we are just seeing how it works by implementing it.

But in sinful it is already implemented.

So now we tried to damp the new instance of our container class which only has simple field services

which will contain an associate array that will consist of their service name and the actual service

class that lives under that name.

So this brings us to how we will identify services which are which are just classes inside our container.

We will identify them by their name name could be a name that you give them arbitrarily or it could

be our class name so that doesnt really matter.

So we have added new function at servies we have a parameter string name and closure.

So you can see that our service has two parameters.

The first one is name and the second is closure.

Closure is the type that you give to represent that you expect the anonymous function to be passed.

So if you would like to remind yourself about the anonymous function you can always go back to the lecture

about them and we will also meet to another functions like the first one is has service so we will check

if the container already contains service with a given name and get service is the actual function that

will or it is actually a method that will let us fetch the specific service that we wanted.

So this method gets service first checks if the container already contains service with a given name.

If it does not it will return.

No but if it could find the service we forgive a name and it would be an instance of closure which means

it is known and must function we will actually codas anonymous function right there and set the value

of the service under that name.

In our services array to the result of that function if that looks complicated to you and I bet it might

then please just bear with me.

You will see how that all works out.

The later part of this lecture.

So now if we have called this closure we can then return the element of the array with the given name

because it is not closure anymore it is now the result of that anonymous function and those anonymous

functions will actually create new instances of classes or lets call them services.

Finally let's also add a method to get services which will return an array with the list of all the

services that we have defined inside the container.

Now its time to see how we can actually add a new service into our service containers.

So lets go back to the index BHB file and call our method of service on our container in stunts and

provide a name for it.

Let's use a format that Jason and now you you can define an anonymous function as you can see it has

to use the container because otherwise their container variable wouldn't be accessible inside this anonymous

function.

And as you can see it simply creates a new instance of the some class.

You can also define a service the same way for example.

And now when you refresh the page you can see the contents of our of our service that it all actually

has two services and there is form adjacent for example and that those services currently contain closures

and now thanks to using their anonymous function to the service instead of some concrete class we can

add another service and call it format and this format service would actually return a different service.

In this case the format.

Jason so this would work the way that when someone is fetching format's service from the container this

would call the anonymous function we defined here and this anonymous function itself has a container

get servies which would in turn return Jason.

Now let's add another servies and call it Serializer

and let's copy the code that we have above for creating a new serializer.

And now you can see that instead of passing the JSM explicitly we can use the container to fetch our

service with a given name.

Let's assume we have a formal service which can be anything.

Depends what we said it to.

So now that we have Serializer service we can also create a service for our index controller.

So let's add another call to service and to the index controller we will pass the serializer service.

Now you can see how we move the logic of creating these classes or services into one place which is

the service container.

So you won't have to manually create these classes in every place whenever they are needed.

And they also want to be couple because whenever something called Meet dependency like Seurat Serializer

or any other class in a matter you can just fetch it from their service container or pass it as a dependency.