18.3 Simple Service Container (Hands-On Coding! )(Section 3: Modern PHP 7 Introduction (OPTIONAL))
So now let's try to create an example and before we see how that works we need to other missing return

statement in service for the format and also fixed that this because this would actually be a container.

And now our code runs and it works but we can see a response because we are not returning it in any

way from the index controller so let's either var dump method call and refresh.

And now you can see that all this is configured properly.

Everything is injected and we have a clear configuration of what are the dependencies of all those classes.

And it's also very easy to configure how our services or objects would work using the service container

because we all have a one place where we configure what should be injected where for example imagine

you would like to convert to SML.

So it's just a simple change to the format service.

And now everything is serialized as SML that's what.

That's because the format service itself is Jassem Elias for any other service that we find her.

But the serializer gets the format and the controller gets the serializer so we don't have to do much

to just change how your application works.