19.1. Service Autowiring Implementation Part 1(Section 3: Modern PHP 7 Introduction (OPTIONAL))
Last time we created a pretty simple yet powerful service container implementation.

The thing that you will use a lot inside Symphonie framework.

And in this lectures we will cover a feature called outer wiring outer wiring is pretty powerful yet

very simple feature that will let us out Tamati Mattioli in Jeg dependencies based on the types of the

constructor arguments.

So let's take a look at the serializer you can see that we have a format interface and for this specific

cases we will implement also a feature of the service container called Elias because it's impossible

to determine which of the classes that implement that specific interface should be injected automatically

wherever we have for example in this controller.

It is much simpler because we have a concrete class specified as a dependency.

Our Serializer so it would be much easier to out to wire it so you won't have to create a definition

for each of those services.

They will be created out of Mattioli.

So let's go back to alliances.

We've added one field which will hold rate.

It's called alliances and now we'll need couple of metals to actually handle this.

Elias is the first one is at Elias.

So Elias as the name suggests is unlike Elias something.

In our case it is Elias to a service name so let's add Elias metalled that will accept two parameters.

One is first is the Elias name and the second one is their service name.

Same as with the services we will need.

Couple of methods to check if they are liars is actually defined.

So lets them has Elias mettled that will check if the given Elias was already defined.

Finally we need a way to fetch our service by Elias instead of by its name.

So let's add another method and call it get Elias.

This method accepts one argument which is the Elias name and it will return servies which we said on

Elias to

as you can see it is calling that get service.

So we have our list of alliances that point to our service and it will try to fetch us service by using

this map of alliances to service names now.

It would also be very useful to specify players when adding new service and you will see why in a bit.

For now does Arban optional Elias parameter to the service method and make it default.

Make it a no by default.

So inside the body of this method we will have to check if Elias is defined.

If it would be no or it wouldn't be provided then this check in the if statement would be false and

if it is provided then we will just add a new Elias to the service.

That is just being created.

Right so we have everything we need.

And now we can go back to the index BHP and import the format interface which is the interface that

all of those free classes are both Jason Exxon-Mobile and the Yarm implemented and that is our first

use case for an alliance.

So as you might have guessed you can't really create a service from an interface.

It has to be a class you cannot instantiate an interface.

So we are adding a new service format we form up interface class name Elias.

So the class Constanta will just give it a fully qualified class name or interface name that doesn't

really matter but it will just give that the class name with an namespace part.

So as you can see now in our container we have five services and we have one alliance the alliance name

is the format interface class name.

Together with the namespace part and it links to the form of service.